\name{cassenade-refits}
\docType{data}

\alias{cassenade-refits}

\title{Dataset: refitting relationships between objects found at the Palaeolithic Cassenade site}
\description{
   Cassenade is a Palaeolithic site excavated... 
}

\usage{data(cassenade)}
\format{
    \itemize{
        \item\code{cassenade.refits} is a 81x10 matrix, describing the refitting relationships between objects at the Cassenade site.
    }
    
}
\references{
    \itemize{
		\item  Discamps, E. 2023. Field database of Cassenade excavations (2012-2013), Nakala, DOI: 10.34847/nkl.e30aie2u.
		\item  Discamps, E., F. Bachellerie, M. Baillet, L. Sitzia. 2019. The Use of Spatial Taphonomy for Interpreting Pleistocene Palimpsests: An Interdisciplinary Approach to the Châtelperronian and Carnivore Occupations at Cassenade (Dordogne, France), PaleoAnthropology, p. 362-388.
		}
	}
\keyword{datasets}

\examples{
data(cassenade)
}

