\name{SEAHORS}
\alias{SEAHORS}

\title{SEAHORS}
\description{Launch SEAHORS}

\usage{
SEAHORS()
}

\details{
This is a launcher function for the `SEAHORS` application.
}
\value{
	Launch the `SEAHORS` Shiny application.
}
\author{Aurelien Royer <aurelien.royer at u-bourgogne.fr>}

\seealso{
	\code{shiny}
    \code{\link[shinythemes]{shinytheme}}
	\code{plotly}
    \code{\link[plotly]{layout}}
	\code{ggplot2}
    \code{\link[mgcv]{gam}}
    \code{\link[cxhull]{cxhull}}
    \code{\link[cxhull]{hullMesh}}
}

\examples{
	\dontrun{
	SEAHORS()
	}
}
